backend/requirements.txt
fastapi == 0.128.0
uvicorn == 0.40.0
sqlalchemy == 2.0.45
psycopg2-binary == 2.9.11
alembic == 1.17.2
python-jose == 3.5.0
passlib[bcrypt]
pgvector == 0.4.2
pydantic == 2.12.5
pydantic-settings
email-validator
httpx
langgraph
openai
pytest>=7.0
typing-extensions>=4.5
python-dotenv